feat(core): implemented service-layer changelog helpers
